Revert "Revert "Rename proxy to stub (#19)" (#20)"
This reverts commit e2e7c392c4c2541451daf4f381ef9577b8e76329.
