+++ src/actors/state.ts
@@ -1,12 +1,12 @@
-import type { create } from "./proxyutil.ts";
 import type { Actor, ActorConstructor } from "./runtime.ts";
 import type { ActorStorage } from "./storage.ts";
+import type { create } from "./stubutil.ts";
 
 export interface ActorStateOptions {
   id: string;
   discriminator?: string;
   storage: ActorStorage;
-  proxy: <TInstance extends Actor>(
+  stub: <TInstance extends Actor>(
     actor: ActorConstructor<TInstance>,
   ) => ReturnType<typeof create<TInstance>>;
 }
@@ -17,13 +17,13 @@ export class ActorState {
   public id: string;
   public discriminator?: string;
   public storage: ActorStorage;
-  public proxy: <TInstance extends Actor>(
+  public stub: <TInstance extends Actor>(
     actor: ActorConstructor<TInstance>,
   ) => ReturnType<typeof create<TInstance>>;
   public initialization: Promise<void> = Promise.resolve();
   constructor(options: ActorStateOptions) {
     this.storage = options.storage;
-    this.proxy = options.proxy;
+    this.stub = options.stub;
     this.id = options.id;
     this.discriminator = options.discriminator;
   }
